+++ backend/src/main/java/com/woowacourse/moragora/service/MeetingService.java
@@ -11,12 +11,12 @@ import com.woowacourse.moragora.entity.Meeting;
 import com.woowacourse.moragora.entity.Participant;
 import com.woowacourse.moragora.entity.Status;
 import com.woowacourse.moragora.entity.user.User;
-import com.woowacourse.moragora.exception.AttendanceNotFoundException;
-import com.woowacourse.moragora.exception.ClosingTimeExcessException;
-import com.woowacourse.moragora.exception.InvalidParticipantException;
-import com.woowacourse.moragora.exception.MeetingNotFoundException;
-import com.woowacourse.moragora.exception.ParticipantNotFoundException;
-import com.woowacourse.moragora.exception.UserNotFoundException;
+import com.woowacourse.moragora.exception.meeting.AttendanceNotFoundException;
+import com.woowacourse.moragora.exception.meeting.ClosingTimeExcessException;
+import com.woowacourse.moragora.exception.meeting.MeetingNotFoundException;
+import com.woowacourse.moragora.exception.participant.InvalidParticipantException;
+import com.woowacourse.moragora.exception.participant.ParticipantNotFoundException;
+import com.woowacourse.moragora.exception.user.UserNotFoundException;
 import com.woowacourse.moragora.repository.AttendanceRepository;
 import com.woowacourse.moragora.repository.MeetingRepository;
 import com.woowacourse.moragora.repository.ParticipantRepository;
