+++ backend/src/main/java/com/woowacourse/moragora/service/UserService.java
@@ -7,7 +7,7 @@ import com.woowacourse.moragora.dto.UsersResponse;
 import com.woowacourse.moragora.entity.user.EncodedPassword;
 import com.woowacourse.moragora.entity.user.User;
 import com.woowacourse.moragora.exception.NoParameterException;
-import com.woowacourse.moragora.exception.UserNotFoundException;
+import com.woowacourse.moragora.exception.user.UserNotFoundException;
 import com.woowacourse.moragora.repository.UserRepository;
 import java.util.List;
 import java.util.stream.Collectors;
