+++ backend/src/test/java/com/woowacourse/moragora/service/MeetingServiceTest.java
@@ -14,11 +14,11 @@ import com.woowacourse.moragora.dto.ParticipantResponse;
 import com.woowacourse.moragora.dto.UserAttendanceRequest;
 import com.woowacourse.moragora.entity.Meeting;
 import com.woowacourse.moragora.entity.Status;
-import com.woowacourse.moragora.exception.ClosingTimeExcessException;
-import com.woowacourse.moragora.exception.InvalidParticipantException;
-import com.woowacourse.moragora.exception.MeetingNotFoundException;
-import com.woowacourse.moragora.exception.ParticipantNotFoundException;
-import com.woowacourse.moragora.exception.UserNotFoundException;
+import com.woowacourse.moragora.exception.meeting.ClosingTimeExcessException;
+import com.woowacourse.moragora.exception.meeting.MeetingNotFoundException;
+import com.woowacourse.moragora.exception.participant.InvalidParticipantException;
+import com.woowacourse.moragora.exception.participant.ParticipantNotFoundException;
+import com.woowacourse.moragora.exception.user.UserNotFoundException;
 import com.woowacourse.moragora.service.closingstrategy.TimeChecker;
 import com.woowacourse.moragora.util.CurrentDateTime;
 import java.time.LocalDate;
